+++ src/caffe/util/im2col.cpp
@@ -60,11 +60,11 @@ template void im2col_cpu<float>(const float* data_im, const int channels,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
     float* data_col);
-template void im2col_cpu<double>(const double* data_im, const int channels,
+template void im2col_cpu<half>(const half* data_im, const int channels,
     const int height, const int width, const int kernel_h, const int kernel_w,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
-    double* data_col);
+    half* data_col);
 
 template <typename Dtype>
 inline void im2col_nd_core_cpu(const Dtype* data_input, const bool im2col,
@@ -153,11 +153,11 @@ template void im2col_nd_cpu<float>(const float* data_im,
     const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
     const int* dilation, float* data_col);
-template void im2col_nd_cpu<double>(const double* data_im,
+template void im2col_nd_cpu<half>(const half* data_im,
     const int num_spatial_axes,
     const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, double* data_col);
+    const int* dilation, half* data_col);
 
 template <typename Dtype>
 void col2im_cpu(const Dtype* data_col, const int channels,
@@ -202,11 +202,11 @@ template void col2im_cpu<float>(const float* data_col, const int channels,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
     float* data_im);
-template void col2im_cpu<double>(const double* data_col, const int channels,
+template void col2im_cpu<half>(const half* data_col, const int channels,
     const int height, const int width, const int kernel_h, const int kernel_w,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
-    double* data_im);
+    half* data_im);
 
 template <typename Dtype>
 void col2im_nd_cpu(const Dtype* data_col, const int num_spatial_axes,
@@ -224,11 +224,11 @@ template void col2im_nd_cpu<float>(const float* data_col,
     const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
     const int* dilation, float* data_im);
-template void col2im_nd_cpu<double>(const double* data_col,
+template void col2im_nd_cpu<half>(const half* data_col,
     const int num_spatial_axes,
     const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, double* data_im);
+    const int* dilation, half* data_im);
 
 
 }  // namespace caffe
